Easy, normal, hard modes
Changed fighter.lua such that an ai can be "easy", "normal" or "hard"
mode by setting the p2.mode = ____ in main.lua to the appropriate
string. This changes the length of time each frame is displayed for the
AI, effectively making its actions faster or slower in animation. This
means parrying reactively is easier in "easy" and "normal" modes, but
very challenging in "hard" mode.